+++ packages/akashic-cli-commons/package.json
@@ -9,7 +9,7 @@
     "build": "tsc -p ./",
     "lint": "tslint --type-check -c ../tslint.json --project ./tsconfig.json",
     "test": "npm run test:jest && npm run lint",
-    "test:jest": "jest"
+    "test:jest": "jest --runInBand"
   },
   "author": "DWANGO Co., Ltd.",
   "license": "MIT",
