+++ packages/akashic-cli-commons/spec/CliConfig/CliConfigurationFileSpec.js
@@ -48,44 +48,3 @@ describe("ConfigurationFile", function () {
 		});
 	});
 });
-
-describe("invalid ConfigurationFile", function () {
-
-	var mockBrokenCAkashicConfigJs = {
-		"akashic.config.js": `
-		var config = {
-			commandOptions: {
-				serve: {
-					port: 30303030
-				}
-			}
-		}
-		module.exports = config;
-		`
-	};
-
-	describe(".read()", function () {
-		beforeEach(function () {
-			mockfs({
-
-			});
-		});
-		afterEach(function () {
-			mockfs.restore();
-		});
-		it("invalid akashic.config.js", function (done) {
-			CliConfigurationFile.read(path.join(process.cwd(), "./akashic.config.js"), (error, config) => {
-				if (error) return done();
-				expect(config).toEqual({
-					commandOptions: {
-						serve: {
-							port: 3000000000
-						}
-					}
-				});
-				done.fail();
-			});
-		});
-	});
-});
-
